+++ FrontEnd/src/routes/AdminRouter.jsx
@@ -3,8 +3,8 @@ import '../pages/AdminPage/AdminGlobal.css';
 import Header from '../components/Admin/Header/Header';
 import Menu from '../pages/AdminPage/Menu/Menu';
 import UserDetail from '../pages/AdminPage/DetailView/UserDetail';
-import UserTable from '../pages/UserTable/UserTable';
-import ProfilesTable from '../pages/UserTable/ProfilesTable';
+import UserTable from '../pages/UserProfilesTable/UserTable';
+import ProfilesTable from '../pages/UserProfilesTable/ProfilesTable';
 import ProfileDetail from '../pages/AdminPage/DetailView/ProfileDetail';
 import css from '../pages/AdminPage/AdminPage.module.css';
 import { Routes, Route } from 'react-router-dom';
